Fix instantiation/visitation of annotated template instantiations
The previous version only considered the targets of annotated type aliases.  It
used an internal facility for ensuring implicit instantiations exist, which also
failed to instantiate members of the class template.  Now the process of
creating an explicit template instantiation definition is emulated instead.

Note that in case an explicit instantiation isn't present in the source, the
specialization declarations are added to the enclosing declaration context of
their template only after the initial parsing, which explains their
position/order in the AST.